announcement/serializers.py
from rest_framework.serializers import ModelSerializer

from announcement.models import Announcement, Feedback


class AnnouncementSerializer(ModelSerializer):
    class Meta:
        model = Announcement
        fields = "__all__"


class FeedbackSerializer(ModelSerializer):
    class Meta:
        model = Feedback
        fields = "__all__"
