pyproject.toml
[tool.poetry]
name = "bulletin-board"
version = "0.1.0"
description = ""
authors = ["andrey-trasov <159887936+andrey-trasov@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
